Add a 'restricted' API
This commit implements cut-down versions of `standalone.js` and `postcssPlugin.js`. The files have been modified to remove functionality that expects a file system.

The restricted API will only accept css and stylelint configs as a string. It can not load css or configs from a file, it can not load plugins. Due to no longer needing a filesystem, the restricted API will run in a browser.